Move database config struct to database package
