--- config.go
@@ -1,29 +1,0 @@
-package zindexer
-
-import (
-	"fmt"
-	"gorm.io/gorm"
-)
-
-type DBConnectionParams struct {
-	User     string
-	Password string
-	Name     string
-	Host     string
-	Port     string
-}
-
-type GraphqlClientParams struct {
-	Host  string
-	Token string
-}
-
-type DBConnectionConfig struct {
-	Gorm *gorm.Config
-}
-
-func (p *DBConnectionParams) GetDSN() (string, error) {
-	return fmt.Sprintf(
-		"user=%s password=%s dbname=%s host=%s port=%s sslmode=disable",
-		p.User, p.Password, p.Name, p.Host, p.Port), nil
-}
