+++ connections/database/interface.go
@@ -1,6 +1,9 @@
 package database
 
-import "gorm.io/gorm"
+import (
+	"fmt"
+	"gorm.io/gorm"
+)
 
 type DBConnection interface {
 	GetDB() *gorm.DB
@@ -13,4 +16,23 @@ type IDBQueryClient interface {
 
 type DBQueryClient struct {
 	Client IDBQueryClient
+}
+
+type DBConnectionParams struct {
+	User     string
+	Password string
+	Name     string
+	Host     string
+	Port     string
+}
+
+type GraphqlClientParams struct {
+	Host  string
+	Token string
+}
+
+func (p *DBConnectionParams) GetDSN() (string, error) {
+	return fmt.Sprintf(
+		"user=%s password=%s dbname=%s host=%s port=%s sslmode=disable",
+		p.User, p.Password, p.Name, p.Host, p.Port), nil
 }
